fix: 룩 이동 단위 테스트가 정상적으로 동작하도록 수정
